requirements.txt
h5py==3.6.0
keras==2.8.0
leafmap==0.20.0
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.8.1
streamlit_option_menu==0.3.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1
